Refactor : 기존 폴더 구조 변경
