+++ src/widgets/layout/Footer/Footer.tsx
